Split store.ts file + added first steps to loading indicator
