+++ frontend/src/js/state/epics/index.ts
@@ -1,0 +1,53 @@
+import { combineEpics } from 'redux-observable';
+
+import {
+    fetchUserEpic,
+    logoutEpic,
+    showToastEpic,
+} from './global';
+
+import {
+    preloadMetadataEpic,
+    addMetadataEpic,
+    editMetadataEpic,
+    deleteMetadataEpic,
+    fetchMetadataEpic,
+    addMetadataTypeEpic,
+    editMetadataTypeEpic,
+    deleteMetadataTypeEpic,
+    fetchMetadataTypesEpic,
+    updateMetadataTypeEpic,
+} from './metadata';
+
+import {
+    fetchContentEpic,
+    addContentEpic,
+    deleteContentEpic,
+    editContentEpic,
+    updateFiltersEpic,
+} from './content';
+
+import * as Utils from './util';
+
+const epics = combineEpics(...[
+    fetchUserEpic,
+    logoutEpic,
+    showToastEpic,
+    preloadMetadataEpic,
+    addMetadataEpic,
+    editMetadataEpic,
+    deleteMetadataEpic,
+    fetchMetadataEpic,
+    addMetadataTypeEpic,
+    editMetadataTypeEpic,
+    deleteMetadataTypeEpic,
+    fetchMetadataTypesEpic,
+    updateMetadataTypeEpic,
+    fetchContentEpic,
+    addContentEpic,
+    deleteContentEpic,
+    editContentEpic,
+    updateFiltersEpic,
+].map(epic => Utils.errorCatcher(epic)));
+
+export default epics;
