+++ frontend/src/js/state/epics/types.d.ts
@@ -1,0 +1,6 @@
+import type { AnyAction } from '@reduxjs/toolkit';
+import type { Epic } from 'redux-observable';
+
+import type { MyState } from '../store';
+
+type MyEpic = Epic<AnyAction, AnyAction, MyState>
